music_platform/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
pydub==0.25.1
python-dotenv==1.0.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
Werkzeug==2.3.7
python-magic==0.4.27 